.history/requirements_20250531105933.txt
Flask==2.3.2
Flask-Cors==4.0.0
py-supabase==1.0.3
bcrypt==4.0.1
streamlit==1.22.0
pandas==2.0.0
PyPDF2==3.0.1
nltk==3.8.1
chardet==5.1.0
openpyxl==3.1.2